Minimal changes to compile with rustc nightly (0.11-pre 33c3edd 2014-05-23)

+++ src/zip/crc32.rs
@@ -71,7 +71,7 @@ pub fn crc32(buf: &[u8]) -> u32 {
     let mut r = 0xffffffffu32;
     for byte in buf.iter() {
         let idx = byte ^ ((r & 0xff) as u8);
-        r = (r >> 8) ^ CRCTABLE[idx];
+        r = (r >> 8) ^ CRCTABLE[idx as uint];
     }
 
     return r ^ 0xffffffffu32;
